Remove activerecord version 6 limit from gemspec
We want to be able to use the `graphql-preload` gem in the `turtle` app
but `turtle` is using Rails 6 and the `graphql-preload` gem requires
`activerecord` be less than version 6. To fix this issue, we forked the
`graphql-preload` gem and are now removing that version upper limit
for `activerecord` so that the forked gem can now be used in `turtle`.

[1157524311697436](https://app.asana.com/0/0/1157524311697436/f)